Create README.md
Added basic premise to the game, and bare bone requirements. Please sign in at the top, add to the blank spots, and put your name next to tasks you wish to help with. Commit to a branch so we can all comment and discuss the changes before merging with the master.